removed specialization of transaction for each protocol version, not needed because could be implemented in a backward compatible way
