+++ server/src/main/java/com/orientechnologies/orient/server/OConnectionBinaryExecutor.java
@@ -76,8 +76,6 @@ import com.orientechnologies.orient.server.network.protocol.binary.*;
 import com.orientechnologies.orient.server.plugin.OServerPlugin;
 import com.orientechnologies.orient.server.tx.OTransactionOptimisticProxy;
 import com.orientechnologies.orient.server.tx.OTransactionOptimisticServer;
-import com.orientechnologies.orient.server.tx.OTransactionOptimisticServer37;
-import com.orientechnologies.orient.server.tx.OTransactionOptimisticServer38;
 
 import java.io.File;
 import java.io.IOException;
@@ -1190,7 +1188,7 @@ public final class OConnectionBinaryExecutor implements OBinaryRequestExecutor {
 
   @Override
   public OBinaryResponse executeBeginTransaction(OBeginTransactionRequest request) {
-    final OTransactionOptimisticServer37 tx = new OTransactionOptimisticServer37(connection.getDatabase(), request.getTxId(),
+    final OTransactionOptimisticServer tx = new OTransactionOptimisticServer(connection.getDatabase(), request.getTxId(),
         request.isUsingLog(), request.getOperations(), request.getIndexChanges());
     try {
       connection.getDatabase().rawBegin(tx);
@@ -1203,7 +1201,7 @@ public final class OConnectionBinaryExecutor implements OBinaryRequestExecutor {
 
   @Override
   public OBinaryResponse executeBeginTransaction38(OBeginTransaction38Request request) {
-    final OTransactionOptimisticServer38 tx = new OTransactionOptimisticServer38(connection.getDatabase(), request.getTxId(),
+    final OTransactionOptimisticServer tx = new OTransactionOptimisticServer(connection.getDatabase(), request.getTxId(),
         request.isUsingLog(), request.getOperations(), request.getIndexChanges());
     try {
       connection.getDatabase().rawBegin(tx);
@@ -1219,7 +1217,7 @@ public final class OConnectionBinaryExecutor implements OBinaryRequestExecutor {
     ODatabaseDocumentInternal database = connection.getDatabase();
     final OTransactionOptimisticServer tx;
     if (request.isHasContent()) {
-      tx = new OTransactionOptimisticServer37(database, request.getTxId(), request.isUsingLog(), request.getOperations(),
+      tx = new OTransactionOptimisticServer(database, request.getTxId(), request.isUsingLog(), request.getOperations(),
           request.getIndexChanges());
       try {
         database.rawBegin(tx);
@@ -1269,7 +1267,7 @@ public final class OConnectionBinaryExecutor implements OBinaryRequestExecutor {
     ODatabaseDocumentInternal database = connection.getDatabase();
     final OTransactionOptimisticServer tx;
     if (request.isHasContent()) {
-      tx = new OTransactionOptimisticServer38(database, request.getTxId(), request.isUsingLog(), request.getOperations(),
+      tx = new OTransactionOptimisticServer(database, request.getTxId(), request.isUsingLog(), request.getOperations(),
           request.getIndexChanges());
       try {
         database.rawBegin(tx);
